@@ -2,27 +2,35 @@
   "name": "@mntm/native-router",
   "version": "0.3.3",
   "type": "module",
-  "main": "dist/index.js",
+  "main": "./dist/index.js",
+  "exports": "./dist/index.js",
   "license": "MIT",
   "sideEffects": false,
   "scripts": {
     "prebuild": "shx rm -rf dist",
     "build": "tsc --build tsconfig.build.json",
-    "test": "jest"
+    "test": "node --experimental-modules --experimental-json-modules --experimental-vm-modules ./node_modules/.bin/jest"
   },
   "devDependencies": {
     "@mntm/eslint-config": "^0.2.0",
-    "@types/jest": "^26.0.23",
-    "@types/jsdom": "^16.2.11",
-    "@types/react": "^17.0.11",
-    "eslint": "^7.28.0",
-    "jest": "^27.0.4",
+    "@types/jest": "^26.0.24",
+    "@types/jsdom": "^16.2.13",
+    "@types/react": "^17.0.15",
+    "esbuild": "^0.12.16",
+    "esbuild-jest": "^0.5.0",
+    "eslint": "^7.31.0",
+    "jest": "^27.0.6",
+    "jest-resolve": "^27.0.6",
+    "jest-ts-webcompat-resolver": "^1.0.0",
     "jsdom": "^16.6.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "shx": "^0.3.3",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.3.2"
+    "ts-jest-resolver": "^1.0.0",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "@mntm/shared": "^2.1.0"
+    "@mntm/shared": "^3.0.0",
+    "mitt": "^3.0.0"
   }
 }
