@@ -1,10 +1,10 @@
 import type { RealHistoryItem, RealHistoryParams, RealHistoryPartial } from './types.js';
 
-import { __dev__ } from '@mntm/shared';
-
 import { realCurrent, realHistory } from './real.js';
 import { popNative } from './native.js';
 import { replace, replacePartial, replacePanel, replaceView, replacePartialParams, replaceParams } from './replace.js';
+
+const __dev__ = process.env.NODE_ENV === 'development';
 
 export const canPop = () => {
   return realHistory.length > 1;
