@@ -1,12 +1,17 @@
 import type { RealHistoryItem, RealHistoryFallback, RealHistoryPartial } from './types.js';
 
-import { findIndex, findLastIndex } from '@mntm/shared';
+import { findIndex, findLastIndex, isPartialEqual } from '@mntm/shared';
 
-import { realHistory, setHistory } from './real.js';
-
+import { afterUpdateHistory } from './history.js';
+import { realCurrent, realHistory, setHistory } from './real.js';
 import { pushPartial } from './push.js';
+import { moveTo } from './move.js';
+import { replacePartial } from './replace.js';
 
 export const changeRoot = (root: string, fallback: Readonly<RealHistoryFallback>) => {
+  const partial = fallback as RealHistoryPartial;
+  const current = realCurrent();
+
   const sameRoot = (item: RealHistoryItem) => {
     return item.root === root;
   };
@@ -16,26 +21,45 @@ export const changeRoot = (root: string, fallback: Readonly<RealHistoryFallback>
     // not found in history
 
     // push new root
-    const partial = fallback as RealHistoryPartial;
     partial.root = root;
     pushPartial(partial);
-  } else {
-    // found in history
-
-    // save ids
-    const ids = realHistory.map((item) => item.id);
-
-    // move all root to start
-    const indexEnd = findLastIndex(realHistory, sameRoot);
-    const savedHistorySlice = realHistory.splice(indexStart, indexEnd - indexStart + 1);
-    realHistory.push(...savedHistorySlice);
-
-    // re-assign all ids
-    realHistory.forEach((item, i) => {
-      item.id = ids[i];
-    });
-
-    // update
-    setHistory(realHistory);
+    return;
   }
+
+  if (current.root === root) {
+    // change to same root
+
+    const start = realHistory[indexStart];
+    if (isPartialEqual(partial, start)) {
+      // started with fallback
+
+      // move to fallback
+      moveTo(indexStart);
+      return;
+    }
+
+    // should move to start and replace with fallback
+    afterUpdateHistory(() => {
+      replacePartial(partial);
+    });
+    moveTo(indexStart);
+  }
+
+  // found in history
+
+  // save ids
+  const ids = realHistory.map((item) => item.id);
+
+  // move all root to start
+  const indexEnd = findLastIndex(realHistory, sameRoot);
+  const savedHistorySlice = realHistory.splice(indexStart, indexEnd - indexStart + 1);
+  realHistory.push(...savedHistorySlice);
+
+  // re-assign all ids
+  realHistory.forEach((item, i) => {
+    item.id = ids[i];
+  });
+
+  // update
+  setHistory(realHistory);
 };
