@@ -22,14 +22,13 @@ export type RealHistoryInit = {
   root: string;
   view: string;
   panel: string;
-
   params?: Optional<RealHistoryParams>;
 };
 
 export type RealHistoryFallback = {
+  root?: string;
   view: string;
   panel: string;
-  params?: Optional<RealHistoryParams>;
 };
 
 export type NativeHistoryItem = {
